neat golang

+++ hello-grpc-go/proto2go.sh
@@ -5,15 +5,28 @@ cd "$(
   pwd -P
 )/" || exit
 export GOROOT=$(go env GOROOT)
-echo "GOROOT=$GOROOT"
-export GOPATH=$HOME/gopath
+export GOPATH=$(go env GOPATH)
+echo "GOROOT=$GOROOT GOPATH=$GOPATH"
 export PATH="$PATH:$GOPATH/bin"
-#
-export GOPROXY=https://mirrors.aliyun.com/goproxy/
-go install google.golang.org/protobuf/cmd/protoc-gen-go@latest
-go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@latest
 
-# proto2go.sh: 18: protoc: not found
+PROTOC_VERSION=$(protoc --version)
+if [ -z "$PROTOC_VERSION" ]; then
+  echo "install protoc firstly, see doc/proto.md"
+  exit 1
+else
+  echo "Protoc version: $PROTOC_VERSION"
+fi
+
+if [ ! -f "$GOPATH/bin/protoc-gen-go.exe" ]; then
+  export GOPROXY=https://mirrors.aliyun.com/goproxy/
+  go install google.golang.org/protobuf/cmd/protoc-gen-go@latest
+  go install google.golang.org/grpc/cmd/protoc-gen-go-grpc@latest
+  go install -v golang.org/x/tools/gopls@latest
+  echo "protoc-gen-go.exe has been installed successfully."
+else
+  echo "protoc-gen-go.exe is already installed."
+fi
+
 echo "generate the messages"
 protoc --go_out=. ./proto/landing.proto
 echo "generate the services"
